% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{browse}
\alias{browse}
\alias{browse.character}
\title{Load a given resource into an HTML browser.}
\usage{
browse(x, ...)

\method{browse}{character}(x, ...)
}
\arguments{
\item{x}{resource to browse}

\item{...}{further arguments to \code{\link[utils]{browseURL}}}
}
\description{
Load a resource into an HTML browser.

S3 generic wrapper replacement of \code{\link[utils]{browseURL}}
}
