% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundeck.R
\name{createProject}
\alias{createProject}
\title{Rundeck - Project Creation}
\usage{
createProject(rundeck, name, config = c(), ...)
}
\arguments{
\item{rundeck}{object of class \code{Rundeck}; see \code{\link{rundeck}}}

\item{name}{project name}

\item{config}{project definition}
}
\value{
nothing
}
\description{
Create a new project.
}
