% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution.R
\name{getExecutionLog}
\alias{getExecutionLog}
\title{Retrieve Rundeck Execution Output Log}
\usage{
getExecutionLog(exec, follow = TRUE, verbose = follow, offset = 0L,
  maxLines = 100L, raw = FALSE, ...)
}
\arguments{
\item{exec}{execution}

\item{follow}{make repeated requests to follow the output of an execution}

\item{verbose}{print every log entry received and a summary at the end}

\item{raw}{return content as returned by API instead of a formatted log}
}
\value{
all received log entries as a single \code{character()} (invisible)
}
\description{
Retrieve Rundeck Execution Output Log
}
