% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundeck.R
\name{getProject}
\alias{getProject}
\title{Rundeck - Get A Project}
\usage{
getProject(rundeck, name, ...)
}
\arguments{
\item{rundeck}{object of class \code{Rundeck}; see \code{\link{rundeck}}}

\item{name}{project name}
}
\description{
Rundeck - Get A Project
}
