% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundeck.R
\name{listProjects}
\alias{listProjects}
\title{Rundeck - Listing Projects}
\usage{
listProjects(rundeck, ...)
}
\arguments{
\item{rundeck}{object of class \code{Rundeck}; see \code{\link{rundeck}}}
}
\value{
vector of projects
}
\description{
List the existing projects on the server.
}
