% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundeck.R
\name{rundeck}
\alias{rundeck}
\title{Connect to a Rundeck server}
\usage{
rundeck(url = "http://localhost", version = 31, user = NULL,
  token = NULL, contextPath = c())
}
\arguments{
\item{url}{TODO}

\item{version}{TODO}

\item{user}{TODO}

\item{token}{TODO}
}
\description{
Connect to a Rundeck server
}
