% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundeck.R
\name{rundeckGET}
\alias{rundeckGET}
\title{Rundeck GET API Request}
\usage{
rundeckGET(rundeck, path, ...)
}
\arguments{
\item{rundeck}{see \code{\link{rundeck}}}

\item{path}{TODO}
}
\description{
Execute a GET API request against Rundeck
}
