% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundeck.R
\name{rundeckHEAD}
\alias{rundeckHEAD}
\title{Rundeck HEAD API Request}
\usage{
rundeckHEAD(rundeck, path, ...)
}
\arguments{
\item{rundeck}{see \code{\link{rundeck}}}

\item{path}{TOOD}

\item{...}{further arguments to \code{\link[httr]{HEAD}}}
}
\description{
Execute a HEAD API request against rundeck
}
