% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundeck.R
\name{rundeckPOST}
\alias{rundeckPOST}
\title{Rundeck POST API Request}
\usage{
rundeckPOST(rundeck, path, body, ...)
}
\arguments{
\item{rundeck}{see \code{\link{rundeck}}}

\item{path}{TODO}

\item{body}{TODO}
}
\description{
Execute a POST API request against Rundeck
}
